ServiceWebTS: Implement iServiceInformation interface
The ServiceWebTS when used on bouquets there was no name information,
this was happening because iServiceInformation interface was missing.

This patch implements iServiceInfo interface in order to provide those
informations.

In order to use servicewebts in bouquets you need to add the following
service reference:

SERVICE 4114:0:1:0:0:0:0:0:0:0:http%3a//URL:NAME